devicestorage/test_enumerate_mobile.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/ChromeUtils.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
// It is a requirement for an sdcard to be loaded, before this test will pass. 
// This is accomplished by adding a --sdcard 20/50...MB to the command line 

SimpleTest.waitForExplicitFinish();
SpecialPowers.addPermission("devicestorage", true, document);
SpecialPowers.addPermission("device-storage", true, document);

mediaTypes = ['music'];
mediaType = 0;
mimeTypes = ['audio/mp3'];
var storage; 
var prefix; 
var file = "abracadabra.mp3"; 
var addedSoFar = 0;
prefix = mediaTypes[0];

function enumerateSuccess(e) {

  if (e.target.result == null) {
    devicestorage_cleanup();
    return;
  }
  
  var filename = e.target.result.name;

  ok(file == e.target.result.name, "Was able to write to deviceStorage on mobile" ); 

  // clean up
  var cleanup = storage.delete(prefix + "/" + file);
  e.target.continue();
}

function handleError(e) {
  ok(false, "handleError was called : " + e.target.error.name);
  devicestorage_cleanup();
}

function addSuccess(e) {
  var cursor = storage.enumerate(prefix);
  cursor.onsuccess = enumerateSuccess;
  cursor.onerror = handleError;
}

function addError(e) {
  ok(false, "addError was called : " + e.target.error.name);
  devicestorage_cleanup();
  SimpleTest.finish();
}

ok(navigator.getDeviceStorage, "Should have getDeviceStorage");
storage = navigator.getDeviceStorage(prefix);

try {
  ok(storage, "Should be able to get the " + prefix + " directory");
  request = storage.addNamed(createRandomBlob(mimeTypes[mediaType]),  prefix + "/" + file);
  request.onsuccess = addSuccess;
  request.onerror = addError;
}
  catch (e) {
    ok(false, "unknown exception called " + e);
    SimpleTest.finish();
  }
</script>
</pre>
</body>
</html>
